# swagger.yaml
openapi: 3.0.0
info:
  description: |
    "This is a Documentation Rest API for Vaccine Booking from capstone project Alterra Academy, Group 46. you can access it here. [http://13.57.43.205/](http://13.57.43.205/) or here [http://ec2-13-57-43-205.us-west-1.compute.amazonaws.com/](http://ec2-13-57-43-205.us-west-1.compute.amazonaws.com/)"
  version: "2.0.0"
  title: VaccineCaps
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: azkazainur@gmail.com
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/Azifaazka/VaccineCaps/2.0.0
  - description: BaseURL
    url: http://13.57.43.205/
...
paths:
...
  /admin/certificate/{id}:
    get:
      tags:
        - Certificate
      summary: Get by id
      operationId: getCertificateByID
      parameters:
        - name: id
          in: path
          description: Use certificate for testing.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Certificate'
            application/xml:
              schema:
                $ref: '#/components/schemas/Certificate'
        '400':
          description: Invalid id supplied
        '404':
          description: Certificates not found
    put:
      tags:
        - Certificate
      summary: Update an existing certificate
      operationId: UpdateCertificates
      parameters:
        - name: id
          in: path
          description: Update
          required: true
          schema:
            type: string
      responses:
        '400':
          description: Invalid ID supplied
        '404':
          description: Certificate not found
        '405':
          description: Validation exception
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Certificate'
        description: Update Certificate object
        required: true
    delete:
      tags:
        - Certificate
      summary: Delete Certificate
      description: This can only be done on Hospitals.
      operationId: DeleteCertificateByID
      parameters:
        - name: id
          in: path
          description: The name that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '400':
          description: Invalid id supplied
        '404':
          description: Hospitals not found
...
components:
  schemas:
...
    Certificate:
      type: object
      properties:
        id:
          type: integer
          format: int64
        sertif:
          type: string
      xml:
        name: Certificate
...